move receive chat message listener to PrivateRoom component
