+++ frontend/src/components/MessagesDisplay/index.tsx
@@ -1,44 +1,15 @@
-import React, { useContext, useEffect, useCallback, useRef } from "react";
-import { SocketContext } from "../../context/socket";
+import React, { useEffect, useRef } from "react";
 import Message from './Message';
-import { useAppSelector, useAppDispatch } from '../../app/hooks';
-import { addMessage } from "../../app/features/roomSlice";
+import { useAppSelector } from '../../app/hooks';
 
 const MessagesDisplay: React.FC = () => {
-  const dispatch = useAppDispatch()
-
-  const socket = useContext(SocketContext)
-  const userId = useAppSelector(state => state.user.socketId)
   const messages = useAppSelector(state => state.room.messages)
 
   const messageEndRef = useRef<null | HTMLDivElement>(null)
 
-  const generateRandomNum = () => {
-    return Math.floor((Math.random() * 10000))
-  }
-
   const scrollToBottom = () => {
     messageEndRef.current!.scrollIntoView({ behavior: 'smooth'})
   }
-
-  const updateMessages = useCallback((message) => {
-    dispatch(addMessage(message))
-  }, [dispatch])
-
-  useEffect(() => {
-    socket.once('receive chat message', ( messageData ) => {
-      console.log('received message')
-      const firstMessageClassName = messages.length === 0 ? 'mt-auto' : ''
-      const userClassName = messageData.userId === userId ? 'peer1Message' : 'peer2Message'
-      const newMessage = {
-        content: messageData.msg,
-        userId: messageData.userId,
-        className: `${firstMessageClassName} ${userClassName}`,
-        id: generateRandomNum()
-      }
-      updateMessages(newMessage)
-    })
-  }, [socket, messages, userId, updateMessages])
 
   useEffect(() => {
     scrollToBottom()
