+++ frontend/src/util/helper.tsx
@@ -1,0 +1,3 @@
+export const generateRandomNum = () => {
+  return Math.floor((Math.random() * 10000))
+}
